E2E test rework is complete.
