+++ test/e2e/bots_tester_test.go
@@ -3,23 +3,10 @@
 package e2e
 
 import (
-	"errors"
-	"fmt"
 	"regexp"
-	"strconv"
-	"strings"
 	"testing"
 
-	"github.com/bwmarrin/discordgo"
-	"github.com/google/uuid"
 	"github.com/sanity-io/litter"
-	"github.com/slack-go/slack"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	"k8s.io/apimachinery/pkg/util/wait"
-
-	"github.com/kubeshop/botkube/pkg/multierror"
-	"github.com/kubeshop/botkube/pkg/utils"
 )
 
 const recentMessagesLimit = 5
@@ -42,46 +29,18 @@ type Channel interface {
 	Identifier() string
 }
 
-type SlackChannel struct {
-	*slack.Channel
-}
-
-func (s *SlackChannel) ID() string {
-	return s.Channel.ID
-}
-func (s *SlackChannel) Name() string {
-	return s.Channel.Name
-}
-func (s *SlackChannel) Identifier() string {
-	return s.Channel.Name
-}
-
-type DiscordChannel struct {
-	*discordgo.Channel
-}
-
-func (s *DiscordChannel) ID() string {
-	return s.Channel.ID
-}
-func (s *DiscordChannel) Name() string {
-	return s.Channel.Name
-}
-func (s *DiscordChannel) Identifier() string {
-	return s.Channel.ID
-}
-
-// BotType to instrument
-type BotType string
+// DriverType to instrument
+type DriverType string
 
 const (
 	// CreateEvent when resource is created
-	SlackBot BotType = "slack"
+	SlackBot DriverType = "slack"
 	// UpdateEvent when resource is updated
-	DiscordBot BotType = "discord"
+	DiscordBot DriverType = "discord"
 )
 
 type BotDriver interface {
-	Type() BotType
+	Type() DriverType
 	InitUsers(t *testing.T)
 	InitChannels(t *testing.T) []func()
 	PostInitialMessage(t *testing.T, channel string)
@@ -97,489 +56,4 @@ type BotDriver interface {
 	SecondChannel() Channel
 	BotUserID() string
 	TesterUserID() string
-}
-
-type slackTester struct {
-	cli           *slack.Client
-	cfg           SlackConfig
-	botUserID     string
-	testerUserID  string
-	channel       Channel
-	secondChannel Channel
-}
-
-func (s *slackTester) Type() BotType {
-	return SlackBot
-}
-
-func (s *slackTester) BotUserID() string {
-	return s.botUserID
-}
-
-func (s *slackTester) TesterUserID() string {
-	return s.testerUserID
-}
-
-func (s *slackTester) Channel() Channel {
-	return s.channel
-}
-
-func (s *slackTester) SecondChannel() Channel {
-	return s.secondChannel
-}
-
-type discordTester struct {
-	cli           *discordgo.Session
-	cfg           DiscordConfig
-	botUserID     string
-	testerUserID  string
-	channel       Channel
-	secondChannel Channel
-}
-
-func (d *discordTester) Type() BotType {
-	return DiscordBot
-}
-
-func (d *discordTester) BotUserID() string {
-	return d.botUserID
-}
-
-func (d *discordTester) TesterUserID() string {
-	return d.testerUserID
-}
-
-func (d *discordTester) Channel() Channel {
-	return d.channel
-}
-
-func (d *discordTester) SecondChannel() Channel {
-	return d.secondChannel
-}
-
-func newSlackTester(slackCfg SlackConfig) (BotDriver, error) {
-	slackCli := slack.New(slackCfg.TesterAppToken)
-	_, err := slackCli.AuthTest()
-	if err != nil {
-		return nil, err
-	}
-
-	return &slackTester{cli: slackCli, cfg: slackCfg}, nil
-}
-
-func (s *slackTester) InitUsers(t *testing.T) {
-	t.Helper()
-	s.botUserID = s.findUserID(t, s.cfg.BotName)
-	s.testerUserID = s.findUserID(t, s.cfg.TesterName)
-}
-
-func (s *slackTester) InitChannels(t *testing.T) []func() {
-	channel, cleanupChannelFn := s.createChannel(t)
-	s.channel = &SlackChannel{Channel: channel}
-
-	secondChannel, cleanupSecondChannelFn := s.createChannel(t)
-	s.secondChannel = &SlackChannel{Channel: secondChannel}
-
-	return []func(){
-		func() { cleanupChannelFn(t) },
-		func() { cleanupSecondChannelFn(t) },
-	}
-}
-
-func newDiscordTester(discordCfg DiscordConfig) (BotDriver, error) {
-	discordCli, err := discordgo.New("Bot " + discordCfg.TesterAppToken)
-	if err != nil {
-		return nil, fmt.Errorf("while creating Discord session: %w", err)
-	}
-	return &discordTester{cli: discordCli, cfg: discordCfg}, nil
-}
-
-func (d *discordTester) InitUsers(t *testing.T) {
-	t.Helper()
-	d.botUserID = d.findUserID(t, d.cfg.BotName)
-	d.testerUserID = d.findUserID(t, d.cfg.TesterName)
-}
-
-func (d *discordTester) InitChannels(t *testing.T) []func() {
-	channel, cleanupChannelFn := d.createChannel(t)
-	d.channel = &DiscordChannel{Channel: channel}
-
-	secondChannel, cleanupSecondChannelFn := d.createChannel(t)
-	d.secondChannel = &DiscordChannel{Channel: secondChannel}
-
-	return []func(){
-		func() { cleanupChannelFn(t) },
-		func() { cleanupSecondChannelFn(t) },
-	}
-}
-
-func (d *discordTester) createChannel(t *testing.T) (*discordgo.Channel, func(t *testing.T)) {
-	t.Helper()
-	randomID := uuid.New()
-	channelName := fmt.Sprintf("%s-%s", channelNamePrefix, randomID.String())
-
-	t.Logf("Creating channel %q...", channelName)
-	channel, err := d.cli.GuildChannelCreate(d.cfg.GuildID, channelName, discordgo.ChannelTypeGuildText)
-	require.NoError(t, err)
-
-	t.Logf("Channel %q (ID: %q) created", channelName, channel.ID)
-
-	cleanupFn := func(t *testing.T) {
-		t.Helper()
-		t.Logf("Deleting channel %q...", channel.Name)
-		// We cannot archive a channel: https://support.discord.com/hc/en-us/community/posts/360042842012-Archive-old-chat-channels
-		_, err := d.cli.ChannelDelete(channel.ID)
-		assert.NoError(t, err)
-	}
-
-	return channel, cleanupFn
-}
-
-func (d *discordTester) PostInitialMessage(t *testing.T, channelID string) {
-	t.Helper()
-	t.Logf("Posting welcome message for channel: %s...", channelID)
-
-	var additionalContextMsg string
-	if d.cfg.AdditionalContextMessage != "" {
-		additionalContextMsg = fmt.Sprintf("%s\n", d.cfg.AdditionalContextMessage)
-	}
-	message := fmt.Sprintf("Hello!\n%s%s", additionalContextMsg, welcomeText)
-	_, err := d.cli.ChannelMessageSend(channelID, message)
-	require.NoError(t, err)
-}
-
-func (d *discordTester) PostMessageToBot(t *testing.T, channel, command string) {
-	message := fmt.Sprintf("<@%s> %s", d.botUserID, command)
-	_, err := d.cli.ChannelMessageSend(channel, message)
-	require.NoError(t, err)
-}
-
-func (d *discordTester) findUserID(t *testing.T, name string) string {
-	t.Log("Getting users...")
-	res, err := d.cli.GuildMembersSearch(d.cfg.GuildID, name, 5)
-	require.NoError(t, err)
-
-	t.Logf("Finding user ID by name %q...", name)
-	for _, m := range res {
-		if !strings.EqualFold(name, m.User.Username) {
-			continue
-		}
-		return m.User.ID
-	}
-
-	return ""
-}
-
-func (d *discordTester) InviteBotToChannel(_ *testing.T, _ string) {
-	// This is not required in Discord.
-	// Bots can't "join" text channels because when you join a server you're already in every text channel.
-	// See: https://stackoverflow.com/questions/60990748/making-discord-bot-join-leave-a-channel
-}
-
-func (d *discordTester) WaitForMessagePostedRecentlyEqual(userID, channelID, expectedMsg string) error {
-	return d.WaitForMessagePosted(userID, channelID, recentMessagesLimit, func(msg string) bool {
-		return strings.EqualFold(msg, expectedMsg)
-	})
-}
-
-func (d *discordTester) WaitForLastMessageContains(userID, channelID, expectedMsgSubstring string) error {
-	return d.WaitForMessagePosted(userID, channelID, 1, func(msg string) bool {
-		return strings.Contains(msg, expectedMsgSubstring)
-	})
-}
-
-func (d *discordTester) WaitForLastMessageEqual(userID, channelID, expectedMsg string) error {
-	return d.WaitForMessagePosted(userID, channelID, 1, func(msg string) bool {
-		return msg == expectedMsg
-	})
-}
-
-func (d *discordTester) WaitForMessagePosted(userID, channelID string, limitMessages int, assertFn MessageAssertion) error {
-	// To always receive message content:
-	// ensure you enable the MESSAGE CONTENT INTENT for the tester bot on the developer portal.
-	// Applications ↦ Settings ↦ Bot ↦ Privileged Gateway Intents
-	// This setting has been enforced from August 31, 2022
-
-	var fetchedMessages []*discordgo.Message
-	var lastErr error
-
-	err := wait.Poll(pollInterval, d.cfg.MessageWaitTimeout, func() (done bool, err error) {
-		messages, err := d.cli.ChannelMessages(channelID, limitMessages, "", "", "")
-		if err != nil {
-			lastErr = err
-			return false, nil
-		}
-
-		fetchedMessages = messages
-		for _, msg := range messages {
-			if msg.Author.ID != userID {
-				continue
-			}
-
-			if !assertFn(msg.Content) {
-				// different message
-				continue
-			}
-
-			return true, nil
-		}
-
-		return false, nil
-	})
-	if lastErr == nil {
-		lastErr = errors.New("message assertion function returned false")
-	}
-	if err != nil {
-		if err == wait.ErrWaitTimeout {
-			return fmt.Errorf("while waiting for condition: last error: %w; fetched messages: %s", lastErr, structDumper.Sdump(fetchedMessages))
-		}
-		return err
-	}
-
-	return nil
-}
-
-func (d *discordTester) WaitForMessagePostedWithAttachment(userID, channelID string, assertFn AttachmentAssertion) error {
-	// To always receive message content:
-	// ensure you enable the MESSAGE CONTENT INTENT for the tester bot on the developer portal.
-	// Applications ↦ Settings ↦ Bot ↦ Privileged Gateway Intents
-	// This setting has been enforced from August 31, 2022
-
-	var fetchedMessages []*discordgo.Message
-	var lastErr error
-
-	err := wait.Poll(pollInterval, d.cfg.MessageWaitTimeout, func() (done bool, err error) {
-		messages, err := d.cli.ChannelMessages(channelID, 1, "", "", "")
-		if err != nil {
-			lastErr = err
-			return false, nil
-		}
-
-		fetchedMessages = messages
-		for _, msg := range messages {
-			if msg.Author.ID != userID {
-				continue
-			}
-
-			if len(msg.Embeds) != 1 {
-				lastErr = err
-				return false, nil
-			}
-
-			embed := msg.Embeds[0]
-
-			if !assertFn(embed.Title, strconv.Itoa(embed.Color), embed.Description) {
-				// different message
-				continue
-			}
-
-			return true, nil
-		}
-
-		return false, nil
-	})
-	if lastErr == nil {
-		lastErr = errors.New("message assertion function returned false")
-	}
-	if err != nil {
-		if err == wait.ErrWaitTimeout {
-			return fmt.Errorf("while waiting for condition: last error: %w; fetched messages: %s", lastErr, structDumper.Sdump(fetchedMessages))
-		}
-		return err
-	}
-
-	return nil
-}
-
-func (d *discordTester) WaitForMessagesPostedOnChannelsWithAttachment(userID string, channelIDs []string, assertFn AttachmentAssertion) error {
-	errs := multierror.New()
-	for _, channelID := range channelIDs {
-		errs = multierror.Append(errs, d.WaitForMessagePostedWithAttachment(userID, channelID, assertFn))
-	}
-
-	return errs.ErrorOrNil()
-}
-
-func (s *slackTester) createChannel(t *testing.T) (*slack.Channel, func(t *testing.T)) {
-	t.Helper()
-	randomID := uuid.New()
-	channelName := fmt.Sprintf("%s-%s", channelNamePrefix, randomID.String())
-
-	t.Logf("Creating channel %q...", channelName)
-	// > There’s no limit to how many unique channels you can have in Slack — go ahead, create as many as you’d like!
-	// Sure, thanks Slack!
-	// Source: https://slack.com/help/articles/201402297-Create-a-channel
-	channel, err := s.cli.CreateConversation(channelName, false)
-	require.NoError(t, err)
-
-	t.Logf("Channel %q (ID: %q) created", channelName, channel.ID)
-
-	cleanupFn := func(t *testing.T) {
-		t.Helper()
-		t.Logf("Archiving channel %q...", channel.Name)
-		// We cannot delete channel: https://stackoverflow.com/questions/46807744/delete-channel-in-slack-api
-		err = s.cli.ArchiveConversation(channel.ID)
-		assert.NoError(t, err)
-	}
-
-	return channel, cleanupFn
-}
-
-func (s *slackTester) PostInitialMessage(t *testing.T, channelName string) {
-	t.Helper()
-	t.Log("Posting welcome message...")
-
-	var additionalContextMsg string
-	if s.cfg.AdditionalContextMessage != "" {
-		additionalContextMsg = fmt.Sprintf("%s\n", s.cfg.AdditionalContextMessage)
-	}
-	message := fmt.Sprintf("Hello!\n%s%s", additionalContextMsg, welcomeText)
-	_, _, err := s.cli.PostMessage(channelName, slack.MsgOptionText(message, false))
-	require.NoError(t, err)
-}
-
-func (s *slackTester) PostMessageToBot(t *testing.T, channel, command string) {
-	message := fmt.Sprintf("<@%s> %s", s.cfg.BotName, command)
-	_, _, err := s.cli.PostMessage(channel, slack.MsgOptionText(message, false))
-	require.NoError(t, err)
-}
-
-func (s *slackTester) findUserID(t *testing.T, name string) string {
-	t.Log("Getting users...")
-	res, err := s.cli.GetUsers()
-	require.NoError(t, err)
-
-	t.Logf("Finding user ID by name %q...", name)
-	for _, u := range res {
-		if u.Name != name {
-			continue
-		}
-		return u.ID
-	}
-
-	return ""
-}
-
-func (s *slackTester) InviteBotToChannel(t *testing.T, channelID string) {
-	t.Logf("Inviting bot with ID %q to the channel with ID %q", s.botUserID, channelID)
-	_, err := s.cli.InviteUsersToConversation(channelID, s.botUserID)
-	require.NoError(t, err)
-}
-
-func (s *slackTester) WaitForMessagePostedRecentlyEqual(userID, channelID, expectedMsg string) error {
-	return s.WaitForMessagePosted(userID, channelID, recentMessagesLimit, func(msg string) bool {
-		return strings.EqualFold(msg, expectedMsg)
-	})
-}
-
-func (s *slackTester) WaitForLastMessageContains(userID, channelID, expectedMsgSubstring string) error {
-	return s.WaitForMessagePosted(userID, channelID, 1, func(msg string) bool {
-		return strings.Contains(msg, expectedMsgSubstring)
-	})
-}
-
-func (s *slackTester) WaitForLastMessageEqual(userID, channelID, expectedMsg string) error {
-	return s.WaitForMessagePosted(userID, channelID, 1, func(msg string) bool {
-		return msg == expectedMsg
-	})
-}
-
-func (s *slackTester) WaitForMessagePosted(userID, channelID string, limitMessages int, assertFn MessageAssertion) error {
-	var fetchedMessages []slack.Message
-	var lastErr error
-	err := wait.Poll(pollInterval, s.cfg.MessageWaitTimeout, func() (done bool, err error) {
-		historyRes, err := s.cli.GetConversationHistory(&slack.GetConversationHistoryParameters{
-			ChannelID: channelID, Limit: limitMessages,
-		})
-		if err != nil {
-			lastErr = err
-			return false, nil
-		}
-
-		fetchedMessages = historyRes.Messages
-		for _, msg := range historyRes.Messages {
-			if msg.User != userID {
-				continue
-			}
-
-			if !assertFn(msg.Text) {
-				// different message
-				continue
-			}
-
-			return true, nil
-		}
-
-		return false, nil
-	})
-	if lastErr == nil {
-		lastErr = errors.New("message assertion function returned false")
-	}
-	if err != nil {
-		if err == wait.ErrWaitTimeout {
-			return fmt.Errorf("while waiting for condition: last error: %w; fetched messages: %s", lastErr, utils.StructDumper().Sdump(fetchedMessages))
-		}
-		return err
-	}
-
-	return nil
-}
-
-func (s *slackTester) WaitForMessagePostedWithAttachment(userID, channelID string, assertFn AttachmentAssertion) error {
-	var fetchedMessages []slack.Message
-	var lastErr error
-	err := wait.Poll(pollInterval, s.cfg.MessageWaitTimeout, func() (done bool, err error) {
-		historyRes, err := s.cli.GetConversationHistory(&slack.GetConversationHistoryParameters{
-			ChannelID: channelID, Limit: 1,
-		})
-		if err != nil {
-			lastErr = err
-			return false, nil
-		}
-
-		fetchedMessages = historyRes.Messages
-		for _, msg := range historyRes.Messages {
-			if msg.User != userID {
-				continue
-			}
-
-			if len(msg.Attachments) != 1 {
-				return false, nil
-			}
-
-			attachment := msg.Attachments[0]
-			if len(attachment.Fields) != 1 {
-				return false, nil
-			}
-
-			if !assertFn(attachment.Title, attachment.Color, attachment.Fields[0].Value) {
-				// different message
-				return false, nil
-			}
-
-			return true, nil
-		}
-
-		return false, nil
-	})
-	if lastErr == nil {
-		lastErr = errors.New("message assertion function returned false")
-	}
-	if err != nil {
-		if err == wait.ErrWaitTimeout {
-			return fmt.Errorf("while waiting for condition: last error: %w; fetched messages: %s", lastErr, structDumper.Sdump(fetchedMessages))
-		}
-		return err
-	}
-
-	return nil
-}
-
-func (s *slackTester) WaitForMessagesPostedOnChannelsWithAttachment(userID string, channelIDs []string, assertFn AttachmentAssertion) error {
-	errs := multierror.New()
-	for _, channelID := range channelIDs {
-		errs = multierror.Append(errs, s.WaitForMessagePostedWithAttachment(userID, channelID, assertFn))
-	}
-
-	return errs.ErrorOrNil()
 }
